refactor(dashboard): split dashboard to application and dashboard
