+++ src/scenes/Dashboard/ApplicationDetail/ApplicationDetail.tsx
