+++ src/scenes/Dashboard/ApplicationFormEditor/components/ApplicationsTable.tsx
