+++ src/scenes/Dashboard/Dashboard.tsx
@@ -1,17 +1,22 @@
-import React from "react";
-import { trpc } from "@/services/trpc";
-import DashboardOrganizer from "@/scenes/Dashboard/DashboardOrganizer/DashboardOrganizer";
-import DashboardHacker from "@/scenes/Dashboard/DashboardHacker/DashboardHacker";
+import { Card } from "@/components/Card";
+import { Heading } from "@/components/Heading";
+import { Button } from "@/components/Button";
+import ApplicationsTable from "@/scenes/Dashboard/ApplicationFormEditor/components/ApplicationsTable";
 
 const Dashboard = () => {
-  const { data } = trpc.userInfo.useQuery();
-  const isOrganizer = Boolean(data?.data?.organizer);
-
-  if (isOrganizer) {
-    return <DashboardOrganizer />;
-  }
-
-  return <DashboardHacker />;
+  return (
+    <Card>
+      <Heading spaceAfter="large" centered>
+        Dashboard Organizer
+      </Heading>
+      <Button
+        label={"Edit application form"}
+        type="buttonLink"
+        href="/dashboard/form-editor"
+      />
+      <ApplicationsTable />
+    </Card>
+  );
 };
 
 export default Dashboard;
