+++ src/scenes/LoginPage/LoginPage.tsx
@@ -29,7 +29,7 @@ const LoginPage = ({ providers }: LoginPageProps) => {
     signIn("credentials", {
       email: data.email,
       password: data.password,
-      callbackUrl: "/dashboard",
+      callbackUrl: "/application",
     });
   };
